@@ -2,7 +2,9 @@
 module.exports = {
   presets: [
     ['@babel/preset-env', { targets: { node: 'current' } }],
-    '@babel/preset-react',
+    ["@babel/preset-react", {
+      "runtime": "automatic"
+    }],
     '@babel/preset-typescript',
   ],
 };
