@@ -1,7 +1,6 @@
 import { fn } from '@storybook/test';
-import { Button, ButtonProps } from './Button';
+import { Button, ButtonProps } from './button';
 import { JSX } from 'react/jsx-runtime';
-import React from 'react';
 
 export const ActionsData = {
   onClick: fn(),
@@ -9,7 +8,7 @@ export const ActionsData = {
 
 export default {
   component: Button,
-  title: 'Button',
+  title: 'Molecules/Button',
   tags: ['autodocs'],
   //👇 Our exports that end in "Data" are not stories.
   excludeStories: /.*Data$/,
